% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{add_cand_length}
\alias{add_cand_length}
\title{Add length of Candidates to Dataframe.}
\usage{
add_cand_length(df)
}
\arguments{
\item{df}{data.frame including cand_tok_1 column with tokenized unigrams.}
}
\value{
Modified data.frame.
}
\description{
New column is added and filled with length of unigrams of candidate.
The length of the sentence in token count is used here.
}
