% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{add_eff_ref_len}
\alias{add_eff_ref_len}
\title{Add Effective Reference Length to data.frame.}
\usage{
add_eff_ref_len(df)
}
\arguments{
\item{df}{data.frame including already added cand_len and tokenized ref token
in column ref_tok_1.}
}
\value{
Modified data.frame.
}
\description{
Add effective reference length to data.frame using the eff_ref_len_atomic
function. Only chose unigrams as described in the Papineni et al., 2002.
}
