% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{bleu}
\alias{bleu}
\title{Compute BLEU.}
\usage{
bleu(ref, cand, n = 4, weights = NA)
}
\arguments{
\item{ref}{List of vectors of references.}

\item{cand}{Vector of candidate sentences.}

\item{n}{Max n-gram size to be considered (default set to 4).}

\item{weights}{Custom weight vector for modified precision for each n-gram
size.}
}
\value{
BLEU-Score for provided corpus.
}
\description{
Compute BLEU for a corpus-based on according to Papineni et al., 2002.
}
\examples{
# Corpus
ref <- list(
  c("The goods cost less than 20 euros.",
    "The merchandise was less than 20 EURO."),
  c("The fee would equal 40\% of the value of the goods...",
    "The fee corresponds with 40 \% of the goods’ value..."),
  c("I am #PRS_ORG# a serious customer and that is why it is not a problem for me.",
    "I am a major client of #PRS_ORG# and thus it is no problem for me."))
cand <- c("The goods cost less than 20 euros.",
    "The fee corresponds to 40\% of the value of the goods....",
    "I'm a #PRS_ORG# major customer so it's not a problem for me.")
bleu_score <- bleu(ref, cand)

# Sentence
ref <- list(
  c("The goods cost less than 20 euros.",
    "The merchandise was less than 20 EURO."))
cand <- c("The goods cost less than 20 euros.")
bleu_score <- bleu(ref, cand)
}
