% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{brevity_pen}
\alias{brevity_pen}
\title{Computes Brevity Penalty}
\usage{
brevity_pen(cand_len, eff_ref_len)
}
\arguments{
\item{cand_len}{Length of the candidate sentence.}

\item{eff_ref_len}{Effective reference length.}
}
\value{
Brevity Penalty.
}
\description{
TODO: Explain brevity penalty.
}
