% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{brevity_penalty}
\alias{brevity_penalty}
\title{Compute Brevity Penalty}
\usage{
brevity_penalty(df)
}
\arguments{
\item{df}{data.frame including already added cand_len and eff_ref_len.}
}
\value{
Modified data.frame.
}
\description{
Compute brevity penalty according to Papineni et al., 2002. Returns 0 iff
candidate length is 0. Provided data.frame must include previously computed
effective reference length column (eff_ref_len) and candidate length
(cand_len).
}
