% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_n_grams}
\alias{combine_n_grams}
\title{Get n-grams from Tokens.}
\usage{
combine_n_grams(ngrams, n)
}
\arguments{
\item{ngrams}{Vector of tokens.}

\item{n}{Parameter to determine range of n-grams.}
}
\value{
Vector of strings entailing requested n-grams.
}
\description{
Collapse tokens to n-grams for list of atomic Strings. Create three lists of
tokens in which sequentially the first token is the subsequently following
token of the first token of the previous list. A data.frame is constructed
based on this list and combined to the requested n-grams.
}
