% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine_n_grams_list}
\alias{combine_n_grams_list}
\title{N-Gram Wrapper.}
\usage{
combine_n_grams_list(ngrams_l, n)
}
\arguments{
\item{ngrams_l}{List of tokens to expand into n-grams.}

\item{n}{Parameter to determine range of n-grams.}
}
\value{
List of vectors containing requested n-grams.
}
\description{
Call function `combine_n_grams` for each entry in list.
}
