% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_df}
\alias{construct_df}
\title{Construct Dataframe.}
\usage{
construct_df(reference, candidate)
}
\arguments{
\item{reference}{List of references.}

\item{candidate}{Vector of candidate strings.}
}
\value{
Dataframe, containing two columns for ref and candidate
in length of input.
}
\description{
Construct Dataframe from references and candidates. References are to be
presented as a list of character vectors while candidates are required to be
character vectors. The reference list and the candidate vector must have the
same length. A Dataframe is constructed from this input.
}
