% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{eff_ref_len}
\alias{eff_ref_len}
\title{Compute BLEU Score}
\usage{
eff_ref_len(cand_len, reference)
}
\arguments{
\item{cand_len}{Length of the candidate sentence.}

\item{reference}{List of reference sentences.}
}
\value{
Effective reference length.
}
\description{
Finds effective reference length, which is necessary to compute the
brevity penalty, that is the closest length to the candidate's length.
}
