% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{eff_ref_len_atomic}
\alias{eff_ref_len_atomic}
\title{Compute Effective Reference Length for Each Pair.}
\usage{
eff_ref_len_atomic(cand_len, reference)
}
\arguments{
\item{cand_len}{length of the candidate sentence.}

\item{reference}{List of reference sentences.}
}
\value{
Effective reference length.
}
\description{
Get all length of references and choose the length which is closest to
the length of the candidate sentence (effective refence length).
}
