% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{mod_prec}
\alias{mod_prec}
\title{Compute Modified Precision for N-Gram.}
\usage{
mod_prec(df_loc)
}
\arguments{
\item{df_loc}{Slice of data depending on n-gram size. Must include candidate
and references in same n-gram size.}
}
\value{
Modified precision for provided corpus.
}
\description{
Compute modified precision based on clipped sums from mod_prec_atomic.
Sum both nominators and denominators and divide fraction.
}
