% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{mod_prec_atomic}
\alias{mod_prec_atomic}
\title{Count and Clip n-grams for a Row.}
\usage{
mod_prec_atomic(ref, cand)
}
\arguments{
\item{ref}{List of reference sentences.}

\item{cand}{A candidate sentence.}
}
\value{
List including nominator and denominator to be divided after
summation.
}
\description{
Occurences are counted and clipped. Clipping is necessary to avoid a higher
candidate token count. Due to the precision-inspired nature, this procedure
consideres the "true positive" occurences.
}
