% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{mod_precision}
\alias{mod_precision}
\title{Compute Modified Precision}
\usage{
mod_precision(candidate, reference, n = 2)
}
\arguments{
\item{candidate}{A candidate sentence (e.g. generated by a model).}

\item{reference}{List of reference sentences.}

\item{n}{n for n-gram}
}
\value{
Modified precision for candidate and reference
}
\description{
Function to compute modified precision. Sum over all n-grams in which the
numerator includes the total number of n-grams from the produced candidate
sentence (clipped at max of the n-gram count from the reference to avoid
larger values than 1). Denominator includes counts of n-grams from reference
sentence.
}
