% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_df}
\alias{process_df}
\title{Process Dataframe.}
\usage{
process_df(df, n = 1)
}
\arguments{
\item{df}{Dataframe, from `construct_df`.}

\item{n}{n-gram level.}
}
\value{
Dataframe with new columns for requested n-gram.
}
\description{
Tokenize input and append as column to df, furthermore add n-grams for each
n > 1 as a separate column. Only tokenizing for unigrams, for all other n > 1
tokens are only expanded to n-grams. Special case for BLEU on sentence level
to avoid treatment of vectors as strings.
}
