% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_ref}
\alias{read_ref}
\title{Read Reference Sentences.}
\usage{
read_ref(files, path = ".")
}
\arguments{
\item{files}{List of filenames of reference files.}

\item{path}{Parent folder for files.}
}
\value{
List including vectors of reference sentences.
}
\description{
Helper function to conveniently read the reference sentences from different
files. The files must be in the format so that each line refers to a
system's output.
}
\examples{
# READ FILES
parent_folder <- "path/to/references"
path_ref_A <- "path2fileA"
path_ref_B <- "path2fileB"
ref <- read_ref(c(path_ref_A, path_ref_B), path = parent_folder)

# READ FILES FROM SAME FOLDER
path_ref_A <- "path2fileA"
path_ref_B <- "path2fileB"
ref <- read_ref(c(path_ref_A, path_ref_B))

# READ FILES WITH ABSOLUTE PATHS
path_ref_A_abs <- "abs/path/path2fileA"
path_ref_B_abs <- "abs/path/path2fileB"
ref <- read_ref(c(path_ref_A_abs, path_ref_B_abs), path = "")
}
