% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{ref_n_gram_count}
\alias{ref_n_gram_count}
\title{Get n-gram Frequency Table}
\usage{
ref_n_gram_count(reference, n = 2)
}
\arguments{
\item{reference}{List of reference sentences.}

\item{n}{Determining n-length of n-gram.}
}
\value{
Frequency table.s
}
\description{
Computes frequency table for n-grams.
}
