% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleu.R
\name{bleu}
\alias{bleu}
\title{Compute BLEU Score}
\usage{
bleu(candidate, reference, weights, N = 4)
}
\arguments{
\item{candidate}{A candidate sentence (e.g. generated by a model).}

\item{reference}{List of reference sentences.}

\item{weights}{List of weights for each n (n-gram). Must be of size N.}

\item{N}{Number n of maximum n-grams.}
}
\value{
BLEU-Score for candidate and reference sentences.
}
\description{
This function computes BLEU (Bilingual Evaluation Understudy) according
to the paper of Papineni et al. 2002. TODO: Write Comprehensive Documentation.
TODO: Write examples
}
