% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_cand}
\alias{read_cand}
\title{Read Candidate Sentences.}
\usage{
read_cand(path)
}
\arguments{
\item{path}{Absolute path to the candidates text file.}
}
\value{
Vector of candidate sentences.
}
\description{
Helper function to conveniently read the candidate sentences from a file.
The file must be of the format in which each line refers to a
system's output. The function is a wrapper for the `readLines` function.
}
\examples{
# READ FILE
cand_name <- "cand"
write(c("ab", "cd"), cand_name)
cand <- read_cand(cand_name)
file.remove(cand_name)
}
